move Entity.js to EntityDialog
