+++ webclient/app/src/app/commons/entityDialog/Entity.js
